Split files into common and games
