--- src/main/java/edu/ntnu/idi/idatt/boardgame/action/Action.java
@@ -1,5 +1,0 @@
-package edu.ntnu.idi.idatt.boardgame.action;
-
-public interface Action {
-  String execute();
-}
