+++ src/main/java/edu/ntnu/idi/idatt/boardgame/common/action/Action.java
@@ -1,0 +1,5 @@
+package edu.ntnu.idi.idatt.boardgame.common.action;
+
+public interface Action {
+  String execute();
+}
