+++ src/main/java/edu/ntnu/idi/idatt/boardgame/common/view/MainView.java
@@ -1,0 +1,3 @@
+package edu.ntnu.idi.idatt.boardgame.common.view;
+
+public class MainView {}
