--- src/main/java/edu/ntnu/idi/idatt/boardgame/domain/board/Snake.java
@@ -1,28 +1,0 @@
-package edu.ntnu.idi.idatt.boardgame.domain.board;
-
-import javafx.scene.paint.Color;
-
-class Snake extends Connector {
-  private final int end;
-  private final Color color = Color.RED;
-
-  public Snake(int start, int length) {
-    super(start);
-    this.end = start - length;
-  }
-
-  @Override
-  public int getEnd() {
-    return end;
-  }
-
-  @Override
-  public Color getColor() {
-    return color;
-  }
-
-  @Override
-  public Object getConnectorType() {
-    return "Snake";
-  }
-}
